+++ src/components/common/IOPopup.vue
@@ -13,7 +13,7 @@
 
 <script setup lang="ts">
 import { ref } from 'vue'
-import { usePopupStore } from '@/stores/popupStore'
+import { usePopupStore } from '@/stores/PopupStore'
 
 const popupShow = ref({ status: false, text: '', type: 'error' })
 const popupStore = usePopupStore()
