+++ src/components/pages/authorization/AuthLogin.vue
@@ -2,7 +2,7 @@
 import { ref } from 'vue'
 import IOButton from '@/components/common/IOButton.vue'
 import IOInput from '@/components/common/IOInput.vue'
-import { useUserStore } from '@/stores/UserStore'
+import { useUserStore } from '@/stores/account/UserStore'
 
 const email = ref('')
 const password = ref('')
