--- src/stores/counter.ts
@@ -1,12 +1,0 @@
-import { defineStore } from 'pinia'
-import { ref, computed } from 'vue'
-
-export const useCounterStore = defineStore('counter', () => {
-  const count = ref(0)
-  const doubleCount = computed(() => count.value * 2)
-  function increment() {
-    count.value++
-  }
-
-  return { count, doubleCount, increment }
-})
