+++ src/views/UiKitView.vue
@@ -23,8 +23,8 @@ import {
   SelectTrigger,
   SelectValue
 } from '@/components/shadcn/ui/select'
-import { useUserStore } from '@/stores/UserStore'
-import { usePopupStore } from '@/stores/popupStore'
+import { useUserStore } from '@/stores/account/UserStore'
+import { usePopupStore } from '@/stores/PopupStore'
 
 const userStore = useUserStore()
 const popupStore = usePopupStore()
